Move invoice document to new created Ginmon specific importer.
