+++ name.abuchen.portfolio.tests/src/name/abuchen/portfolio/datatransfer/pdf/ginmon/FeeInvoice01.txt
