+++ name.abuchen.portfolio/src/name/abuchen/portfolio/datatransfer/pdf/DABPDFExtractor.java
@@ -29,7 +29,6 @@ public class DABPDFExtractor extends AbstractPDFExtractor
         addBankIdentifier("DAB Bank");
         addBankIdentifier("DAB bank AG");
         addBankIdentifier("BNP Paribas S.A. Niederlassung Deutschland");
-        addBankIdentifier("Ginmon GmbH");
 
         addBuySellTransaction();
         addDividendTransaction();
@@ -37,7 +36,6 @@ public class DABPDFExtractor extends AbstractPDFExtractor
         addTaxAdjustmentTransaction();
         addAdvanceTaxTransaction();
         addAccountStatementTransaction();
-        addNonImportableTransaction();
     }
 
     @Override
@@ -1590,63 +1588,4 @@ public class DABPDFExtractor extends AbstractPDFExtractor
                         .assign((t, v) -> processFeeEntries(t, v, type));
     }
 
-    private void addNonImportableTransaction()
-    {
-        final DocumentType type = new DocumentType("(Gebührenabrechnung)", //
-                        documentContext -> documentContext //
-                        // @formatter:off
-                        // Rechnungsdatum / Invoice date: 31.10.2024
-                        // Rechnungsnummer /Invoice no: 00000002024101
-                        // @formatter:on
-                                        .section("date") //
-                                        .match("^Rechnungsdatum / Invoice date: (?<date>[\\d]{2}\\.[\\d]{2}\\.[\\d]{4})$") //
-                                        .assign((ctx, v) -> {
-                                            ctx.put("date", v.get("date"));
-                                        }));
-
-        this.addDocumentTyp(type);
-
-        Transaction<AccountTransaction> pdfTransaction = new Transaction<>();
-
-        Block firstRelevantLine = new Block("^(Gebührenabrechnung) (.*)$");
-        type.addBlock(firstRelevantLine);
-        firstRelevantLine.set(pdfTransaction);
-
-        pdfTransaction //
-
-                        .subject(() -> {
-                            AccountTransaction accountTransaction = new AccountTransaction();
-                            accountTransaction.setType(AccountTransaction.Type.FEES);
-                            return accountTransaction;
-                        })
-
-                        // @formatter:off
-                        // Gebührenabrechnung Oktober
-                        // Invoice October
-                        // Zeitraum vom 01.10.2024 – 31.10.2024
-                        // Period of 01.10.2024 – 31.10.2024
-                        // Position Berechungsbasis Gebühr Betrag
-                        // Calculation Fee Amount
-                        // Grundgebühr 57.60 € 0.7500% p.a. 0.04 €
-                        // @formatter:on
-                        .section("amount", "currency") //
-                        .documentContext("date") //
-                        .match("^Rechnungsbetrag (?<amount>[\\.,\\d]+) (?<currency>\\p{Sc})$") //
-                        .assign((t, v) -> {
-                            t.setDateTime(asDate(v.get("date")));
-                            t.setCurrencyCode(asCurrencyCode(v.get("currency")));
-                            t.setAmount(asAmount(v.get("amount"), "en", "US"));
-                            v.getTransactionContext().put(FAILURE,
-                                            Messages.MsgErrorTransactionAlternativeDocumentRequired);
-                        })
-
-                        .wrap((t, ctx) -> {
-                            TransactionItem item = new TransactionItem(t);
-
-                            if (ctx.getString(FAILURE) != null)
-                                item.setFailureMessage(ctx.getString(FAILURE));
-
-                            return item;
-                        });
-    }
 }
